Don't search parent scope for a local if it's a placeholder
For a snippet like:

  proc do |(a, b)|
    proc do |(c, d)|
    end
  end

it would produce a tree like (comments are mine)

  [:script,
  [:call,
    nil,
    :proc,
    [:arglist,
    [:iter,
      [:args, [:masgn, [:array, [:lasgn, :a], [:lasgn, :b]], [:lvar, :"_:1"]]], # (1)
      [:call,
      nil,
      :proc,
      [:arglist,
        [:iter,
        [:args,
          [:masgn, [:array, [:lasgn, :c], [:lasgn, :d]], [:lvar, :"_:1"]]], # (2)
        [:nil]]]]]]]]

When the compiler encounters the `lvar` in "(2)", it would search the
parent scope for a variable named "_:1" since it's not in the current
scope. However, there's a variable with such name in the parent scope in
"(1)", so the compiler thinks we are referencing another variable, which
will lead to variables `c` and `d` in the inner block assigned the
values of `a` and `b`.

Fixes rubinius/rubinius#3544